requirements.txt
agno
streamlit
requests
beautifulsoup4
fpdf
python-dotenv